Modified Fader Components to be EnumMaps
Mute Buttons, Scribbles and Faders are all attached to each other, so adding a generic 'Faders' key and using EnumMaps makes more sense for easy access.

In addition, the way I was using vecs before could crash the daemon if the XML had buttons, scribbles or fader tags out of order (trying to insert ID 2 before ID 1 had been added), so this also fixes that.

+++ profile/src/components/sample.rs
@@ -257,18 +257,13 @@ impl SampleStack {
 #[derive(Debug)]
 struct Track {
     track: String,
-    start_position: f32,
-    end_position: f32,
+    start_position: u8,
+    end_position: u8,
     normalized_gain: f64,
 }
 
 impl Track {
-    pub fn new(
-        track: String,
-        start_position: f32,
-        end_position: f32,
-        normalized_gain: f64,
-    ) -> Self {
+    pub fn new(track: String, start_position: u8, end_position: u8, normalized_gain: f64) -> Self {
         Self {
             track,
             start_position,
